requirements.txt
pyglet
cocos2d
pylint